requirements.txt
Flask==3.0.0
fpdf2==2.7.6
Werkzeug==3.0.1

